Restructure socket hooks to proper location
- Move socket hooks from mirrord/layer/src/hooks/socket/mod.rs to mirrord/layer/src/socket/hooks.rs
- Update socket.rs to declare hooks module with pub(super) mod hooks
- Update Go layer imports to use socket::hooks::* instead of hooks::socket::*
- Remove socket module declaration from hooks.rs
- This matches the proper module structure and fixes compilation errors

The socket detour functions are now properly located within the socket module where they belong.
