+++ mirrord/layer/src/go/linux_x64.rs
@@ -5,10 +5,7 @@ use nix::errno::Errno;
 use tracing::trace;
 
 use crate::{
-    close_detour,
-    file::hooks::*,
-    hooks::{HookManager, socket::*},
-    macros::hook_symbol,
+    close_detour, file::hooks::*, hooks::HookManager, macros::hook_symbol, socket::hooks::*,
 };
 /*
  * Reference for which syscalls are managed by the handlers:
